+++ lfs/gitscanner_tree.go
@@ -1,24 +1,13 @@
 package lfs
 
 import (
-	"bufio"
-	"bytes"
 	"fmt"
-	"io"
 	"io/ioutil"
-	"strconv"
-	"strings"
 
 	"github.com/git-lfs/git-lfs/config"
 	"github.com/git-lfs/git-lfs/filepathfilter"
 	"github.com/git-lfs/git-lfs/git"
 )
-
-// An entry from ls-tree or rev-list including a blob sha and tree path
-type TreeBlob struct {
-	Sha1     string
-	Filename string
-}
 
 func runScanTree(cb GitScannerFoundPointer, ref string, filter *filepathfilter.Filter, gitEnv, osEnv config.Environment) error {
 	// We don't use the nameMap approach here since that's imprecise when >1 file
@@ -59,7 +48,7 @@ func catFileBatchTree(treeblobs *TreeBlobChannelWrapper, gitEnv, osEnv config.En
 	go func() {
 		hasNext := true
 		for t := range treeblobs.Results {
-			hasNext = scanner.Scan(t.Sha1)
+			hasNext = scanner.Scan(t.Oid)
 
 			if p := scanner.Pointer(); p != nil {
 				p.Name = t.Filename
@@ -107,13 +96,13 @@ func lsTreeBlobs(ref string, filter *filepathfilter.Filter) (*TreeBlobChannelWra
 
 	cmd.Stdin.Close()
 
-	blobs := make(chan TreeBlob, chanBufSize)
+	blobs := make(chan git.TreeBlob, chanBufSize)
 	errchan := make(chan error, 1)
 
 	go func() {
-		scanner := newLsTreeScanner(cmd.Stdout)
+		scanner := git.NewLsTreeScanner(cmd.Stdout)
 		for scanner.Scan() {
-			if t := scanner.TreeBlob(); t != nil && filter.Allows(t.Filename) {
+			if t := scanner.TreeBlob(); t != nil && t.Size < blobSizeCutoff && filter.Allows(t.Filename) {
 				blobs <- *t
 			}
 		}
@@ -128,78 +117,4 @@ func lsTreeBlobs(ref string, filter *filepathfilter.Filter) (*TreeBlobChannelWra
 	}()
 
 	return NewTreeBlobChannelWrapper(blobs, errchan), nil
-}
-
-type lsTreeScanner struct {
-	s    *bufio.Scanner
-	tree *TreeBlob
-}
-
-func newLsTreeScanner(r io.Reader) *lsTreeScanner {
-	s := bufio.NewScanner(r)
-	s.Split(scanNullLines)
-	return &lsTreeScanner{s: s}
-}
-
-func (s *lsTreeScanner) TreeBlob() *TreeBlob {
-	return s.tree
-}
-
-func (s *lsTreeScanner) Err() error {
-	return nil
-}
-
-func (s *lsTreeScanner) Scan() bool {
-	t, hasNext := s.next()
-	s.tree = t
-	return hasNext
-}
-
-func (s *lsTreeScanner) next() (*TreeBlob, bool) {
-	hasNext := s.s.Scan()
-	line := s.s.Text()
-	parts := strings.SplitN(line, "\t", 2)
-	if len(parts) < 2 {
-		return nil, hasNext
-	}
-
-	attrs := strings.SplitN(parts[0], " ", 4)
-	if len(attrs) < 4 {
-		return nil, hasNext
-	}
-
-	if attrs[1] != "blob" {
-		return nil, hasNext
-	}
-
-	sz, err := strconv.ParseInt(strings.TrimSpace(attrs[3]), 10, 64)
-	if err != nil {
-		return nil, hasNext
-	}
-
-	if sz < blobSizeCutoff {
-		sha1 := attrs[2]
-		filename := parts[1]
-		return &TreeBlob{Sha1: sha1, Filename: filename}, hasNext
-	}
-	return nil, hasNext
-}
-
-func scanNullLines(data []byte, atEOF bool) (advance int, token []byte, err error) {
-	if atEOF && len(data) == 0 {
-		return 0, nil, nil
-	}
-
-	if i := bytes.IndexByte(data, '\000'); i >= 0 {
-		// We have a full null-terminated line.
-		return i + 1, data[0:i], nil
-	}
-
-	// If we're at EOF, we have a final, non-terminated line. Return it.
-	if atEOF {
-		return len(data), data, nil
-	}
-
-	// Request more data.
-	return 0, nil, nil
 }
